requirements.txt
Flask~=1.1.2
pip~=21.1.1
cryptography~=3.3.1
py~=1.10.0
lxml~=4.6.2
setuptools~=56.2.0
future~=0.18.2
python-dotenv~=0.17.1
rsa~=4.7.2
alembic~=1.6.2
SQLAlchemy~=1.4.15
uuid~=1.30
Werkzeug~=1.0.1
PyJWT~=2.1.0
marshmallow~=3.12.1